% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{fastmanh}
\alias{fastmanh}
\title{Plots an optimised Manhattan plot}
\usage{
fastmanh(
  data,
  signif = 5e-08,
  build = 38,
  maxpeaks = 30,
  no_distance = FALSE,
  no_annot = FALSE
)
}
\arguments{
\item{data}{A data.table containing columns \code{"chr"}, \code{"pos"}, \code{"p"}}

\item{signif}{significance threshold to use. Defaults to 5e-8}

\item{build}{genomic build. Default 38. Can be 37.}

\item{maxpeaks}{The first \code{"maxpeaks"} signals in terms of p-value will be displayed. Increasing this may clutter the plot.}

\item{no_distance}{do not display distances in plot, default FALSE.}

\item{no_annot}{do not compute annotations using VEP. default FALSE.}
}
\value{
A list with \code{"retm"} and \code{"peaks"} attributes. The former is the optimised manhattan object, the latter is the peak object.
}
\description{
A traditional Manhattan plot will contain a point for each variant. This is unnecessary, since the scale of the genome makes many points overlap in a human-sized plot. \code{"fastmanh"} addresses this by thinning the data frame to only visible points.
}
\examples{
\donttest{
library(data.table)
mygwas=read.table("GWAS.GEMMA.assoc.txt.gz")
setnames(mygwas, c("ps", "p_score"), c("pos", "p"))
fastmanh(mygwas)
}
}
