% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{fastqq}
\alias{fastqq}
\title{Display a QQ-plot of a p-value vector from a GWAS}
\usage{
fastqq(pvalue, X_GRID = 800, Y_GRID = 800)
}
\arguments{
\item{pvalue}{A vector of p-values}

\item{X_GRID}{The horizontal resolution of the grid used to simplify the plot. 800 is good for most cases.}

\item{Y_GRID}{The vertical resolution of the grid used to simplify the plot. 800 is good for most cases.}
}
\description{
Display a QQ-plot of a p-value vector from a GWAS
}
\examples{
\donttest{
library(data.table)
mygwas=read.table("GWAS.GEMMA.assoc.txt.gz")
qqplot(mygwas$P_SCORE)
}
}
