% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{get_peaks_to_annotate}
\alias{get_peaks_to_annotate}
\title{Get peaks to annotate}
\usage{
get_peaks_to_annotate(manhattan_object, assoc, signif = 5e-08, build = 38)
}
\arguments{
\item{manhattan_object}{Output object of `mhp` function}

\item{assoc}{data.table with columns chr, pos, a1, a2, p, and af}

\item{signif}{Significance threshold (default: 5e-8)}

\item{build}{Genome assembly number (38 or 37)}
}
\value{
A data.table with columns: plotpos, ploty, p, build
}
\description{
Get peaks to annotate
}
