% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{mhp}
\alias{mhp}
\title{Make manhattan object}
\usage{
mhp(chr, ps, p, X_RES = 2000, Y_RES = 1000, signif = 5e-08)
}
\arguments{
\item{chr}{Chromosome column of data.table}

\item{ps}{Position column of data.table}

\item{p}{P-value column of data.table}

\item{signif}{Significance threshold (default: 5e-8)}
}
\value{
A list with 'newcoords', 'posdict', and 'labpos' attributes
}
\description{
Make manhattan object
}
