% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{query_ensembl_gene_overlap}
\alias{query_ensembl_gene_overlap}
\title{Query Ensembl for overlapping genes in a given genomic region}
\usage{
query_ensembl_gene_overlap(chr, start, end, build = 38)
}
\arguments{
\item{chr}{Chromosome}

\item{start}{Start genomic position}

\item{end}{End genomic position}

\item{build}{The assembly to query}
}
\value{
data.frame with query results. Empty data.frame returned if no overlap
}
\description{
Query Ensembl for overlapping genes in a given genomic region
}
\examples{
query_ensembl_gene_overlap(9, 5073770, 5073770)
query_ensembl_gene_overlap(9, 4973770, 5173770)
}
